#1845, move OKPacket out of backend
