+++ sharding-proxy/src/main/java/org/apache/shardingsphere/shardingproxy/backend/text/transaction/SkipBackendHandler.java
@@ -19,8 +19,8 @@ package org.apache.shardingsphere.shardingproxy.backend.text.transaction;
 
 import org.apache.shardingsphere.shardingproxy.backend.ResultPacket;
 import org.apache.shardingsphere.shardingproxy.backend.text.TextProtocolBackendHandler;
+import org.apache.shardingsphere.shardingproxy.transport.common.packet.generic.DatabaseSuccessPacket;
 import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.command.CommandResponsePackets;
-import org.apache.shardingsphere.shardingproxy.transport.mysql.packet.generic.OKPacket;
 
 /**
  * Skip backend handler.
@@ -31,7 +31,7 @@ public final class SkipBackendHandler implements TextProtocolBackendHandler {
     
     @Override
     public CommandResponsePackets execute() {
-        return new CommandResponsePackets(new OKPacket(1, 0L, 0L));
+        return new CommandResponsePackets(new DatabaseSuccessPacket(1, 0L, 0L));
     }
     
     @Override
